requirements.txt
aiofiles==0.8.0
arrow==1.2.2
dynaconf==3.1.8
fastapi==0.78.0
gunicorn==20.1.0
matplotlib==3.5.2
numpy==1.22.3
Pillow==9.1.1
python-multipart==0.0.5
uvicorn[standard]==0.17.6
